39 Python3 Web框架/3 FastAPI/project/config/config-nacos/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
nacos-sdk-python
